+++ src/stream/async.rs
@@ -1,22 +1,25 @@
 use std::io;
 
-use consts::{NlFamily, NlType};
-use err::NlError;
-use socket::{tokio, NlSocket};
-use Nl;
+use crate::{
+    consts::{NlFamily, NlType},
+    err::NlError,
+    socket::{tokio, NlSocket},
+    utils::{U32BitFlag, U32Bitmask},
+    Nl,
+};
 
 /// Builder for netlink stream
 pub struct NetlinkStreamConnector {
     sock: NlSocket,
-    mcast_ids: Vec<u32>,
+    mcast_ids: U32Bitmask,
 }
 
 impl NetlinkStreamConnector {
     /// Create builder for netlink stream
     pub fn new(proto: NlFamily, pid: Option<u32>) -> Result<Self, io::Error> {
         Ok(NetlinkStreamConnector {
-            sock: NlSocket::connect(proto, pid, None)?,
-            mcast_ids: Vec::new(),
+            sock: NlSocket::connect(proto, pid, U32Bitmask::empty())?,
+            mcast_ids: U32Bitmask::empty(),
         })
     }
 
@@ -26,8 +29,19 @@ impl NetlinkStreamConnector {
         family: &str,
         mcast_group: &str,
     ) -> Result<&mut Self, NlError> {
-        self.mcast_ids
-            .push(self.sock.resolve_nl_mcast_group(family, mcast_group)?);
+        let group_num = self.sock.resolve_nl_mcast_group(family, mcast_group)?;
+        self.mcast_ids += U32BitFlag::new(group_num);
+        Ok(self)
+    }
+
+    /// Remove a multicast group subscription by family and group name
+    pub fn remove_mcast_group(
+        &mut self,
+        family: &str,
+        mcast_group: &str,
+    ) -> Result<&mut Self, NlError> {
+        let group_num = self.sock.resolve_nl_mcast_group(family, mcast_group)?;
+        self.mcast_ids -= U32BitFlag::new(group_num);
         Ok(self)
     }
 
@@ -48,8 +62,14 @@ impl NetlinkStreamConnector {
 
 /// High level API for receiving a stream of netlink messages
 pub struct NetlinkStream<T, P> {
+    // Dead code allowed for CI
+    #[allow(dead_code)]
     sock: tokio::NlSocket<T, P>,
+    // Dead code allowed for CI
+    #[allow(dead_code)]
     pid: Option<u32>,
+    // Dead code allowed for CI
+    #[allow(dead_code)]
     seq: Option<u32>,
 }
 
@@ -62,7 +82,7 @@ where
     pub fn new(
         proto: NlFamily,
         pid: Option<u32>,
-        groups: Option<Vec<u32>>,
+        groups: U32Bitmask,
     ) -> Result<Self, io::Error> {
         Ok(NetlinkStream {
             sock: tokio::NlSocket::new(NlSocket::connect(proto, pid, groups)?)?,